requirements.txt
flask==3.0.3
google-api-python-client==2.140.0
google-auth==2.34.0